feat(ui5-toast): WebComponent is called Toast now
CSS rule for the "BottomCenter" placement is now removed
and added to the default styles of the ".ui5-toast-root-parent".

"placement" attribute is now removed from the "ui5-toast-root" div in
Toast.hbs file.

Incorrect indentations in the hbs file are now fixed.

Since tag added in the documentation of the Toast web component.
Additionally minor grammatical corrections are made.
Description is added for the "show" method of the Toast.

Sample HTML files are simplified to use id's instead of CSS classes
for the button click handlers. This way we have more readable show
trigger of the Toast inside the click handlers of the buttons.